+++ src/color/theme.rs
@@ -38,6 +38,7 @@ pub struct Permission {
     pub exec: Color,
     pub exec_sticky: Color,
     pub no_access: Color,
+    pub octal: Color,
 }
 
 #[derive(Debug, Deserialize, PartialEq)]
@@ -132,6 +133,7 @@ impl Default for Permission {
             exec: Color::DarkRed,
             exec_sticky: Color::AnsiValue(5),
             no_access: Color::AnsiValue(245), // Grey
+            octal: Color::AnsiValue(6),
         }
     }
 }