+++ src/flags/permission.rs
@@ -1,0 +1,167 @@
+//! This module defines the [PermissionFlag]. To set it up from [ArgMatches], a [Config] and its
+//! [Default] value, use its [configure_from](Configurable::configure_from) method.
+
+use super::Configurable;
+
+use crate::config_file::Config;
+
+use clap::ArgMatches;
+use serde::Deserialize;
+
+/// The flag showing which file permissions units to use.
+#[derive(Clone, Debug, Copy, PartialEq, Eq, Deserialize)]
+#[serde(rename_all = "kebab-case")]
+pub enum PermissionFlag {
+    /// The variant to show file permissions in rwx format
+    Rwx,
+    /// The variant to show file permissions in octal format
+    Octal,
+}
+
+impl PermissionFlag {
+    fn from_str(value: &str) -> Option<Self> {
+        match value {
+            "rwx" => Some(Self::Rwx),
+            "octal" => Some(Self::Octal),
+            _ => {
+                panic!(
+                    "Permissions can only be one of rwx or octal, but got {}.",
+                    value
+                );
+            }
+        }
+    }
+}
+
+impl Configurable<Self> for PermissionFlag {
+    /// Get a potential `PermissionFlag` variant from [ArgMatches].
+    ///
+    /// If any of the "rwx" or "octal" arguments is passed, the corresponding
+    /// `PermissionFlag` variant is returned in a [Some]. If neither of them is passed,
+    /// this returns [None].
+    /// Sets permissions to rwx if classic flag is enabled.
+    fn from_arg_matches(matches: &ArgMatches) -> Option<Self> {
+        if matches.is_present("classic") {
+            return Some(Self::Rwx);
+        } else if matches.occurrences_of("permission") > 0 {
+            if let Some(permissions) = matches.values_of("permission")?.last() {
+                return Self::from_str(permissions);
+            }
+        }
+        None
+    }
+
+    /// Get a potential `PermissionFlag` variant from a [Config].
+    ///
+    /// If the `Config::permissions` has value and is one of "rwx" or "octal",
+    /// this returns the corresponding `PermissionFlag` variant in a [Some].
+    /// Otherwise this returns [None].
+    /// Sets permissions to rwx if classic flag is enabled.
+    fn from_config(config: &Config) -> Option<Self> {
+        if let Some(true) = config.classic {
+            Some(Self::Rwx)
+        } else {
+            config.permission
+        }
+    }
+}
+
+/// The default value for `PermissionFlag` is [PermissionFlag::Default].
+impl Default for PermissionFlag {
+    fn default() -> Self {
+        Self::Rwx
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use super::PermissionFlag;
+
+    use crate::app;
+    use crate::config_file::Config;
+    use crate::flags::Configurable;
+
+    #[test]
+    fn test_default() {
+        assert_eq!(PermissionFlag::Rwx, PermissionFlag::default());
+    }
+
+    #[test]
+    fn test_from_arg_matches_none() {
+        let argv = vec!["lsd"];
+        let matches = app::build().get_matches_from_safe(argv).unwrap();
+        assert_eq!(None, PermissionFlag::from_arg_matches(&matches));
+    }
+
+    #[test]
+    fn test_from_arg_matches_default() {
+        let argv = vec!["lsd", "--permission", "rwx"];
+        let matches = app::build().get_matches_from_safe(argv).unwrap();
+        assert_eq!(
+            Some(PermissionFlag::Rwx),
+            PermissionFlag::from_arg_matches(&matches)
+        );
+    }
+
+    #[test]
+    fn test_from_arg_matches_short() {
+        let args = vec!["lsd", "--permission", "octal"];
+        let matches = app::build().get_matches_from_safe(args).unwrap();
+        assert_eq!(
+            Some(PermissionFlag::Octal),
+            PermissionFlag::from_arg_matches(&matches)
+        );
+    }
+
+    #[test]
+    #[should_panic]
+    fn test_from_arg_matches_unknown() {
+        let args = vec!["lsd", "--permission", "unknown"];
+        let _ = app::build().get_matches_from_safe(args).unwrap();
+    }
+    #[test]
+    fn test_from_arg_matches_permissions_multi() {
+        let args = vec!["lsd", "--permission", "octal", "--permission", "rwx"];
+        let matches = app::build().get_matches_from_safe(args).unwrap();
+        assert_eq!(
+            Some(PermissionFlag::Rwx),
+            PermissionFlag::from_arg_matches(&matches)
+        );
+    }
+
+    #[test]
+    fn test_from_arg_matches_permissions_classic() {
+        let args = vec!["lsd", "--permission", "rwx", "--classic"];
+        let matches = app::build().get_matches_from_safe(args).unwrap();
+        assert_eq!(
+            Some(PermissionFlag::Rwx),
+            PermissionFlag::from_arg_matches(&matches)
+        );
+    }
+
+    #[test]
+    fn test_from_config_none() {
+        assert_eq!(None, PermissionFlag::from_config(&Config::with_none()));
+    }
+
+    #[test]
+    fn test_from_config_rwx() {
+        let mut c = Config::with_none();
+        c.permission = Some(PermissionFlag::Rwx);
+        assert_eq!(Some(PermissionFlag::Rwx), PermissionFlag::from_config(&c));
+    }
+
+    #[test]
+    fn test_from_config_octal() {
+        let mut c = Config::with_none();
+        c.permission = Some(PermissionFlag::Octal);
+        assert_eq!(Some(PermissionFlag::Octal), PermissionFlag::from_config(&c));
+    }
+
+    #[test]
+    fn test_from_config_classic_mode() {
+        let mut c = Config::with_none();
+        c.classic = Some(true);
+        assert_eq!(Some(PermissionFlag::Rwx), PermissionFlag::from_config(&c));
+    }
+}