+++ src/meta/permissions.rs
@@ -1,4 +1,5 @@
 use crate::color::{ColoredString, Colors, Elem};
+use crate::flags::{Flags, PermissionFlag};
 use std::fs::Metadata;
 
 #[derive(Debug, PartialEq, Eq, Copy, Clone)]
@@ -54,7 +55,11 @@ impl<'a> From<&'a Metadata> for Permissions {
 }
 
 impl Permissions {
-    pub fn render(&self, colors: &Colors) -> ColoredString {
+    fn bits_to_octal(r: bool, w: bool, x: bool) -> u8 {
+        (r as u8) * 4 + (w as u8) * 2 + (x as u8)
+    }
+
+    pub fn render(&self, colors: &Colors, flags: &Flags) -> ColoredString {
         let bit = |bit, chr: &'static str, elem: &Elem| {
             if bit {
                 colors.colorize(String::from(chr), elem)
@@ -63,35 +68,53 @@ impl Permissions {
             }
         };
 
-        let strings: &[ColoredString] = &[
-            // User permissions
-            bit(self.user_read, "r", &Elem::Read),
-            bit(self.user_write, "w", &Elem::Write),
-            match (self.user_execute, self.setuid) {
-                (false, false) => colors.colorize(String::from("-"), &Elem::NoAccess),
-                (true, false) => colors.colorize(String::from("x"), &Elem::Exec),
-                (false, true) => colors.colorize(String::from("S"), &Elem::ExecSticky),
-                (true, true) => colors.colorize(String::from("s"), &Elem::ExecSticky),
-            },
-            // Group permissions
-            bit(self.group_read, "r", &Elem::Read),
-            bit(self.group_write, "w", &Elem::Write),
-            match (self.group_execute, self.setgid) {
-                (false, false) => colors.colorize(String::from("-"), &Elem::NoAccess),
-                (true, false) => colors.colorize(String::from("x"), &Elem::Exec),
-                (false, true) => colors.colorize(String::from("S"), &Elem::ExecSticky),
-                (true, true) => colors.colorize(String::from("s"), &Elem::ExecSticky),
-            },
-            // Other permissions
-            bit(self.other_read, "r", &Elem::Read),
-            bit(self.other_write, "w", &Elem::Write),
-            match (self.other_execute, self.sticky) {
-                (false, false) => colors.colorize(String::from("-"), &Elem::NoAccess),
-                (true, false) => colors.colorize(String::from("x"), &Elem::Exec),
-                (false, true) => colors.colorize(String::from("T"), &Elem::ExecSticky),
-                (true, true) => colors.colorize(String::from("t"), &Elem::ExecSticky),
-            },
-        ];
+        let strings = match flags.permission {
+            PermissionFlag::Rwx => vec![
+                // User permissions
+                bit(self.user_read, "r", &Elem::Read),
+                bit(self.user_write, "w", &Elem::Write),
+                match (self.user_execute, self.setuid) {
+                    (false, false) => colors.colorize(String::from("-"), &Elem::NoAccess),
+                    (true, false) => colors.colorize(String::from("x"), &Elem::Exec),
+                    (false, true) => colors.colorize(String::from("S"), &Elem::ExecSticky),
+                    (true, true) => colors.colorize(String::from("s"), &Elem::ExecSticky),
+                },
+                // Group permissions
+                bit(self.group_read, "r", &Elem::Read),
+                bit(self.group_write, "w", &Elem::Write),
+                match (self.group_execute, self.setgid) {
+                    (false, false) => colors.colorize(String::from("-"), &Elem::NoAccess),
+                    (true, false) => colors.colorize(String::from("x"), &Elem::Exec),
+                    (false, true) => colors.colorize(String::from("S"), &Elem::ExecSticky),
+                    (true, true) => colors.colorize(String::from("s"), &Elem::ExecSticky),
+                },
+                // Other permissions
+                bit(self.other_read, "r", &Elem::Read),
+                bit(self.other_write, "w", &Elem::Write),
+                match (self.other_execute, self.sticky) {
+                    (false, false) => colors.colorize(String::from("-"), &Elem::NoAccess),
+                    (true, false) => colors.colorize(String::from("x"), &Elem::Exec),
+                    (false, true) => colors.colorize(String::from("T"), &Elem::ExecSticky),
+                    (true, true) => colors.colorize(String::from("t"), &Elem::ExecSticky),
+                },
+            ],
+            PermissionFlag::Octal => {
+                let octal_sticky = Self::bits_to_octal(self.setuid, self.setgid, self.sticky);
+                let octal_user =
+                    Self::bits_to_octal(self.user_read, self.user_write, self.user_execute);
+                let octal_group =
+                    Self::bits_to_octal(self.group_read, self.group_write, self.group_execute);
+                let octal_other =
+                    Self::bits_to_octal(self.other_read, self.other_write, self.other_execute);
+                vec![colors.colorize(
+                    format!(
+                        "{}{}{}{}",
+                        octal_sticky, octal_user, octal_group, octal_other
+                    ),
+                    &Elem::Octal,
+                )]
+            }
+        };
 
         let res = strings
             .iter()
@@ -129,4 +152,134 @@ mod modes {
     pub const STICKY: Mode = libc::S_ISVTX as Mode;
     pub const SETGID: Mode = libc::S_ISGID as Mode;
     pub const SETUID: Mode = libc::S_ISUID as Mode;
+}
+
+#[cfg(unix)]
+#[cfg(test)]
+mod test {
+    use super::Flags;
+    use super::{PermissionFlag, Permissions};
+    use crate::color::{Colors, ThemeOption};
+    use std::fs;
+    use std::fs::File;
+    use std::os::unix::fs::PermissionsExt;
+    use tempfile::tempdir;
+
+    #[test]
+    pub fn permission_rwx() {
+        let tmp_dir = tempdir().expect("failed to create temp dir");
+
+        // Create the file;
+        let file_path = tmp_dir.path().join("file.txt");
+        File::create(&file_path).expect("failed to create file");
+        fs::set_permissions(&file_path, fs::Permissions::from_mode(0o655))
+            .expect("unable to set permissions to file");
+        let meta = file_path.metadata().expect("failed to get meta");
+
+        let colors = Colors::new(ThemeOption::NoColor);
+        let perms = Permissions::from(&meta);
+
+        assert_eq!(
+            "rw-r-xr-x",
+            perms.render(&colors, &Flags::default()).content()
+        );
+    }
+
+    #[test]
+    pub fn permission_rwx2() {
+        let tmp_dir = tempdir().expect("failed to create temp dir");
+
+        // Create the file;
+        let file_path = tmp_dir.path().join("file.txt");
+        File::create(&file_path).expect("failed to create file");
+        fs::set_permissions(&file_path, fs::Permissions::from_mode(0o777))
+            .expect("unable to set permissions to file");
+        let meta = file_path.metadata().expect("failed to get meta");
+
+        let colors = Colors::new(ThemeOption::NoColor);
+        let perms = Permissions::from(&meta);
+
+        assert_eq!(
+            "rwxrwxrwx",
+            perms.render(&colors, &Flags::default()).content()
+        );
+    }
+
+    #[test]
+    pub fn permission_rwx_sticky() {
+        let tmp_dir = tempdir().expect("failed to create temp dir");
+
+        // Create the file;
+        let file_path = tmp_dir.path().join("file.txt");
+        File::create(&file_path).expect("failed to create file");
+        fs::set_permissions(&file_path, fs::Permissions::from_mode(0o1777))
+            .expect("unable to set permissions to file");
+
+        let meta = file_path.metadata().expect("failed to get meta");
+
+        let colors = Colors::new(ThemeOption::NoColor);
+        let mut flags = Flags::default();
+        flags.permission = PermissionFlag::Rwx;
+        let perms = Permissions::from(&meta);
+
+        assert_eq!("rwxrwxrwt", perms.render(&colors, &flags).content());
+    }
+
+    #[test]
+    pub fn permission_octal() {
+        let tmp_dir = tempdir().expect("failed to create temp dir");
+
+        // Create the file;
+        let file_path = tmp_dir.path().join("file.txt");
+        File::create(&file_path).expect("failed to create file");
+        fs::set_permissions(&file_path, fs::Permissions::from_mode(0o655))
+            .expect("unable to set permissions to file");
+        let meta = file_path.metadata().expect("failed to get meta");
+
+        let colors = Colors::new(ThemeOption::NoColor);
+        let mut flags = Flags::default();
+        flags.permission = PermissionFlag::Octal;
+        let perms = Permissions::from(&meta);
+
+        assert_eq!("0655", perms.render(&colors, &flags).content());
+    }
+
+    #[test]
+    pub fn permission_octal2() {
+        let tmp_dir = tempdir().expect("failed to create temp dir");
+
+        // Create the file;
+        let file_path = tmp_dir.path().join("file.txt");
+        File::create(&file_path).expect("failed to create file");
+        fs::set_permissions(&file_path, fs::Permissions::from_mode(0o777))
+            .expect("unable to set permissions to file");
+        let meta = file_path.metadata().expect("failed to get meta");
+
+        let colors = Colors::new(ThemeOption::NoColor);
+        let mut flags = Flags::default();
+        flags.permission = PermissionFlag::Octal;
+        let perms = Permissions::from(&meta);
+
+        assert_eq!("0777", perms.render(&colors, &flags).content());
+    }
+
+    #[test]
+    pub fn permission_octal_sticky() {
+        let tmp_dir = tempdir().expect("failed to create temp dir");
+
+        // Create the file;
+        let file_path = tmp_dir.path().join("file.txt");
+        File::create(&file_path).expect("failed to create file");
+        fs::set_permissions(&file_path, fs::Permissions::from_mode(0o1777))
+            .expect("unable to set permissions to file");
+
+        let meta = file_path.metadata().expect("failed to get meta");
+
+        let colors = Colors::new(ThemeOption::NoColor);
+        let mut flags = Flags::default();
+        flags.permission = PermissionFlag::Octal;
+        let perms = Permissions::from(&meta);
+
+        assert_eq!("1777", perms.render(&colors, &flags).content());
+    }
 }